UnexceptionalIO

Description

When you've caught all the exceptions that can be handled safely, this is what you're left with.

runEitherIO . fromIO ≡ id

Documentation

data UIO a

IO without any non-error, synchronous exceptions

class Unexceptional m where

Polymorphic base without any non-error, synchronous exceptions

Minimal complete definition

liftUIO

Methods

liftUIO :: UIO a -> m a

fromIO :: IO a -> UIO (Either SomeException a)

Catch any non-error, synchronous exceptions in an IO action

runUIO :: UIO a -> IO a

Re-embed UIO into IO

runEitherIO :: Exception e => UIO (Either e a) -> IO a

Re-embed UIO and possible exception back into IO

Unsafe entry points

fromIO' :: Exception e => IO a -> UIO (Either e a)

You promise that e covers all non-error, synchronous exceptions thrown by this IO action

This function is partial if you lie

unsafeFromIO :: IO a -> UIO a

You promise there are no exceptions thrown by this IO action

Utilities

syncIO :: IO a -> IO (Either SomeException a)

Catch all exceptions, except for asynchronous exceptions found in base